@@ -1,24 +1,26 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <AssemblyName>DiegoG.WebWatcherLibrary</AssemblyName>
     <RootNamespace>DiegoG.WebWatcher</RootNamespace>
     <Authors>DiegoG1019</Authors>
     <Description>A library to create extensions for https://github.com/DiegoG1019/WebWatcher</Description>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.0.12</Version>
+    <Version>1.0.13</Version>
+	<Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="DiegoG.RssFeedParser" Version="1.1.0" />
-    <PackageReference Include="DiegoG.TelegramBot" Version="1.7.3" />
-    <PackageReference Include="DiegoG.Utilities.Basic" Version="1.0.2" />
+    <PackageReference Include="DiegoG.TelegramBot" Version="1.7.10" />
+    <PackageReference Include="DiegoG.Utilities.Basic" Version="1.5.2" />
     <PackageReference Include="DiegoG.Utilities.Collections" Version="1.0.0" />
     <PackageReference Include="DiegoG.Utilities.Reflection" Version="1.1.2" />
-    <PackageReference Include="DiegoG.Utilities.Serialization" Version="1.0.0" />
-    <PackageReference Include="HtmlAgilityPack" Version="1.11.35" />
-    <PackageReference Include="Humanizer.Core" Version="2.11.10" />
+    <PackageReference Include="DiegoG.Utilities.Serialization" Version="1.0.1" />
+    <PackageReference Include="DiegoG.Utilities.Settings" Version="1.0.6" />
+    <PackageReference Include="HtmlAgilityPack" Version="1.11.49" />
+    <PackageReference Include="Humanizer.Core" Version="2.14.1" />
   </ItemGroup>
 
 </Project>
