@@ -6,31 +6,30 @@ using System.Text;
 using System.Threading.Tasks;
 using DiegoG.Utilities.Settings;
 
-namespace DiegoG.WebWatcher
+namespace DiegoG.WebWatcher;
+
+public class WatcherSettings : ISettings
 {
-    public class WatcherSettings : ISettings
+    public event PropertyChangedEventHandler? PropertyChanged;
+    public string SettingsType => "DiegoG.WebWatcher.Secrets";
+    public ulong Version => 5;
+    public Serilog.Events.LogEventLevel FileLogEventLevel { get; set; } = Serilog.Events.LogEventLevel.Debug;
+    public Serilog.Events.LogEventLevel BotLogEventLevel { get; set; } = Serilog.Events.LogEventLevel.Information;
+    public Serilog.Events.LogEventLevel ConsoleLogEventLevel { get; set; } = Serilog.Events.LogEventLevel.Debug;
+    public string? BotAPIKey { get; set; }
+
+    public string? VersionName { get; set; } = "";
+
+    public long LogChatId { get; set; } = 0;
+
+    public IDictionary<string, bool> EnableList { get; set; } = new Dictionary<string, bool>();
+
+    public IDictionary<string, bool> ExtensionsEnable { get; set; } = new Dictionary<string, bool>();
+
+    public WatcherSettings()
     {
-        public event PropertyChangedEventHandler? PropertyChanged;
-        public string SettingsType => "DiegoG.WebWatcher.Secrets";
-        public ulong Version => 5;
-        public Serilog.Events.LogEventLevel FileLogEventLevel { get; set; } = Serilog.Events.LogEventLevel.Debug;
-        public Serilog.Events.LogEventLevel BotLogEventLevel { get; set; } = Serilog.Events.LogEventLevel.Information;
-        public Serilog.Events.LogEventLevel ConsoleLogEventLevel { get; set; } = Serilog.Events.LogEventLevel.Debug;
-        public string? BotAPIKey { get; set; }
-
-        public string? VersionName { get; set; } = "";
-
-        public long LogChatId { get; set; } = 0;
-
-        public IDictionary<string, bool> EnableList { get; set; } = new Dictionary<string, bool>();
-
-        public IDictionary<string, bool> ExtensionsEnable { get; set; } = new Dictionary<string, bool>();
-
-        public WatcherSettings()
-        {
 #if DEBUG
-            FileLogEventLevel = Serilog.Events.LogEventLevel.Verbose;
+        FileLogEventLevel = Serilog.Events.LogEventLevel.Verbose;
 #endif
-        }
     }
 }
